refactor: include InterestPeriod in AccrualCycles (#1933)
* fix: Fix calculation of number of days in an interest cycle

* refactor: use InterestPeriod in AccrualCycle

* fix: make bats test pass in ci

---------

+++ core/credit/src/history/entry.rs
@@ -45,7 +45,7 @@ pub struct DisbursalExecuted {
 pub struct InterestAccrualsPosted {
     pub cents: UsdCents,
     pub recorded_at: DateTime<Utc>,
-    pub days: u16,
+    pub days: u32,
     pub tx_id: LedgerTxId,
 }
